refactor: 유저의 입력을 받는 컨테이너와 결과를 나타내는 컨테이너 분리
